pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-notes"
version = "0.2.0"
description = "로컬 폴더에 떨어진 오디오 파일을 Whisper로 전사하고 Claude로 상세·요약 마크다운을 생성"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20240930",
    "imageio-ffmpeg>=0.5.1",
    "python-dotenv>=1.0.0",
    "watchdog>=4.0.0",
]

[project.scripts]
voice-notes = "src.main:main"

[tool.setuptools]
packages = ["src"]
